requirements.txt
fastapi==0.111.0
uvicorn==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
opencv-python==4.10.0.84
numpy==2.1.1
Pillow==10.4.0
regex==2024.9.11
pyyaml==6.0.2
loguru==0.7.2
watchdog==5.0.2
pytesseract==0.3.13
